Added FlySpell, refactored SpellCaster a bit +more
Added in the FlySpell, it seems to work properly.  It will likely break
other mods that also allow flight if some flight item is being used when
the player casts the FlySpell.  The problem lies in the fact that you
cannot determine if another mod grants flight while you have granted
flight.  Ideally flight would be tracked by instances of the buff, and
this might be accomplishable at some point in the future if we instead
apply a potion-effect of flight...maybe?  Possibly will look into it.

The FlySpell required a bit of a refactor of the SpellCaster system, and
it it is becoming a bit hacky.  It will likely need to be rewritten to
accomodate the various changes to the way it works from the initial
conception.  But, it's workable for now.

Switched from eclipse to IDEA.

Updated forge version to 1.8-11.14.2.1435.